refactor(core): move `findUsersByRoleName()` to library
