+++ packages/core/src/queries/user.ts
@@ -12,8 +12,8 @@ import type { Search } from '#src/utils/search.js';
 import { buildConditionsFromSearch } from '#src/utils/search.js';
 
 // TODO: @sijie remove this
-import { findRoleByRoleName, findRolesByRoleIds } from './roles.js';
-import { findUsersRolesByRoleId, findUsersRolesByUserId } from './users-roles.js';
+import { findRolesByRoleIds } from './roles.js';
+import { findUsersRolesByUserId } from './users-roles.js';
 
 const { table, fields } = convertToIdentifiers(Users);
 
@@ -216,22 +216,6 @@ export const createUserQueries = (pool: CommonQueryMethods) => {
       group by date(${fields.createdAt})
     `);
 
-  const findUsersByRoleName = async (roleName: string) => {
-    const role = await findRoleByRoleName(roleName);
-
-    if (!role) {
-      return [];
-    }
-
-    const usersRoles = await findUsersRolesByRoleId(role.id);
-
-    if (usersRoles.length === 0) {
-      return [];
-    }
-
-    return findUsersByIds(usersRoles.map(({ userId }) => userId));
-  };
-
   return {
     findUserByUsername,
     findUserByEmail,
@@ -251,7 +235,6 @@ export const createUserQueries = (pool: CommonQueryMethods) => {
     deleteUserIdentity,
     hasActiveUsers,
     getDailyNewUserCountsByTimeInterval,
-    findUsersByRoleName,
   };
 };
 
@@ -275,5 +258,4 @@ export const {
   deleteUserIdentity,
   hasActiveUsers,
   getDailyNewUserCountsByTimeInterval,
-  findUsersByRoleName,
 } = createUserQueries(envSet.pool);
